Crate abxml

Library that decodes the binary documents contained on an APK (both resources.arsc and binary XMLs). It exposes also structures to query this binary files on a structured way. For example, it's possible to check which chunks of data a document contains, and perform specific queries depending on the type of chunk.

Modules

apk
High level abstraction to easy the extraction to file system of APKs

chunks
Structs to represent chunks and iterate them

decoder
Structs which contains the read data from binary documents

encoder
Exports the decoded binary XMLs to string XMLs

errors
Representation of library errors

model
Representations of logical structures found on android binary files

visitor
Collection of visitors that are fed from chunk iterator

Constants

STR_ARSC
Contents of android's resources.arsc