Module abxml::model

Representations of logical structures found on android binary files

Modules

builder

owned

Structs

Element

ElementContainer

Tag

Enums

Value
Represents a value on the binary documents. It is formed by a type and a 32 bits payload. The payloads are interpreted depending on the type.

Traits

AttributeTrait

Configuration

Identifier

Library

LibraryBuilder

NamespaceEnd

NamespaceStart

Package

Resources

StringTable

TableType

TagEnd

TagStart
Trait that represents a XML tag start

TypeSpec

Type Definitions

Entries

Namespaces