Trait abxml::model::AttributeTrait

pub trait AttributeTrait {
    fn get_namespace(&self) -> Result<u32, Error>;

    fn get_name(&self) -> Result<u32, Error>;

    fn get_class(&self) -> Result<u32, Error>;

    fn get_resource_value(&self) -> Result<u32, Error>;

    fn get_data(&self) -> Result<u32, Error>;

    fn get_value(&self) -> Result<Value, Error> { ... }
}

Required Methods

fn get_namespace(&self) -> Result<u32, Error>

Return the namespace index. If there is no namespace, it will return 0xFFFF_FFFF

fn get_name(&self) -> Result<u32, Error>

Returns the index of the attribute on the string table

fn get_class(&self) -> Result<u32, Error>

Returns the ¿class?

fn get_resource_value(&self) -> Result<u32, Error>

Returns the data type (see Values)

fn get_data(&self) -> Result<u32, Error>

Returns the value (see Values)

Provided Methods

fn get_value(&self) -> Result<Value, Error>

Creates a Value depending on the data type and data value