Struct abxml::visitor::Executor

pub struct Executor;

Methods to decode a binary resource.arsc file or a binary xml file

Methods

impl Executor

fn arsc<'a, V: ChunkVisitor<'a>>(buffer: &'a [u8],
visitor: &mut V)
-> Result<()>

Given a valid resources.arsc file contents, it will call to the proper methods on the given visitor.

fn xml<'a, V: ChunkVisitor<'a>>(cursor: Cursor<&'a [u8]>,
visitor: &mut V)
-> Result<()>

Given a valid binary XML file contents, it will call to the proper methods on the given visitor.