Trait abxml::model::TagStart

pub trait TagStart {
    type Attribute: AttributeTrait;
    fn get_line(&self) -> Result<u32, Error>;

    fn get_field1(&self) -> Result<u32, Error>;

    fn get_namespace_index(&self) -> Result<u32, Error>;

    fn get_element_name_index(&self) -> Result<u32, Error>;

    fn get_field2(&self) -> Result<u32, Error>;

    fn get_attributes_amount(&self) -> Result<u32, Error>;

    fn get_class(&self) -> Result<u32, Error>;

    fn get_attribute(&self, index: u32) -> Result<Self::Attribute, Error>;
}

Trait that represents a XML tag start

Associated Types

type Attribute: AttributeTrait

Type of the attributes

Loading content...
Required methods

fn get_line(&self) -> Result<u32, Error>

Return the ¿line in which the tag appear?

fn get_field1(&self) -> Result<u32, Error>

Return the content of the unknown field1

fn get_namespace_index(&self) -> Result<u32, Error>

Return the namespace index. If there is no namespace, it will return 0xFFFF_FFFF

fn get_element_name_index(&self) -> Result<u32, Error>

Returns the index of the tag name on the string table

fn get_field2(&self) -> Result<u32, Error>

Return the content of the unknown field1

fn get_attributes_amount(&self) -> Result<u32, Error>

Return the amount of attributes this tag contains

fn get_class(&self) -> Result<u32, Error>

Returns the ¿class?

fn get_attribute(&self, index: u32) -> Result<Self::Attribute, Error>

Returns the attribute on the index position or error if it is greater than get_attributes_amount

Loading content...